Move UprobesCallstackManager and UprobesFunctionCallManager to their own files
Move their tests as well, from UprobesUnwindingVisitorTest.cpp.
